+++ Civi/Api4/GeometryCollectionType.php
@@ -10,4 +10,12 @@ namespace Civi\Api4;
  */
 class GeometryCollectionType extends Generic\DAOEntity {
 
+  public static function permissions() {
+    return [
+      'create' => [['administer geometry', 'administer civicrm']],
+      'delete' => [['administer geometry', 'administer civicrm']],
+      'default' => ['access geometry'],
+    ];
+  }
+
 }